requirement.txt
Django==4.2
djangorestframework==3.14.0
psutil==5.9.5
scikit-learn==1.3.0
joblib==1.3.2